Move utils/math to math lib
