+++ math/tests/math_utils.spec.cpp
@@ -1,4 +1,4 @@
-#include <rawrbox/utils/math.hpp>
+#include <rawrbox/math/utils/math.hpp>
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers.hpp>
